Refactor the VectorApprox to a single location
