+++ dartsim/src/WorldFeatures_TEST.cc
@@ -47,29 +47,7 @@ using namespace gz;
 
 using TestEnginePtr = physics::Engine3dPtr<TestFeatureList>;
 using TestWorldPtr = physics::World3dPtr<TestFeatureList>;
-
-// A predicate-formatter for asserting that two vectors are approximately equal.
-class AssertVectorApprox
-{
-  public: explicit AssertVectorApprox(double _tol = 1e-6) : tol(_tol)
-  {
-  }
-
-  public: ::testing::AssertionResult operator()(
-              const char *_mExpr, const char *_nExpr, Eigen::Vector3d _m,
-              Eigen::Vector3d _n)
-  {
-    if (gz::physics::test::Equal(_m, _n, this->tol))
-      return ::testing::AssertionSuccess();
-
-    return ::testing::AssertionFailure()
-           << _mExpr << " and " << _nExpr << " ([" << _m.transpose()
-           << "] and [" << _n.transpose() << "]"
-           << ") are not equal";
-  }
-
-  private: double tol;
-};
+using AssertVectorApprox = physics::test::AssertVectorApprox;
 
 //////////////////////////////////////////////////
 class WorldFeaturesFixture : public ::testing::Test
