+++ test/common_test/shape_features.cc
@@ -66,28 +66,7 @@ class ShapeFeaturesTest:
   public: gz::plugin::Loader loader;
 };
 
-// A predicate-formatter for asserting that two vectors are approximately equal.
-class AssertVectorApprox
-{
-  public: explicit AssertVectorApprox(double _tol = 1e-6) : tol(_tol)
-  {
-  }
-
-  public: ::testing::AssertionResult operator()(
-              const char *_mExpr, const char *_nExpr, Eigen::Vector3d _m,
-              Eigen::Vector3d _n)
-  {
-    if (gz::physics::test::Equal(_m, _n, this->tol))
-      return ::testing::AssertionSuccess();
-
-    return ::testing::AssertionFailure()
-           << _mExpr << " and " << _nExpr << " ([" << _m.transpose()
-           << "] and [" << _n.transpose() << "]"
-           << ") are not equal";
-  }
-
-  private: double tol;
-};
+using AssertVectorApprox = gz::physics::test::AssertVectorApprox;
 
 struct ShapeFeaturesList : gz::physics::FeatureList<
   gz::physics::AttachFixedJointFeature,
